Fix issues in PR #326
There is a CMake language error in PR #326.
If the variable X is already defined (to TRUE or to FALSE or to anything) when calling
cmake_language(CALL "check_${Y}_compiler_flag" "-Wall" "X" ) for any Y, then the
test is skipped, and the variable remains unchanged.

As a result, the function check_and_set_compiler_option would take the flags of the C
compiler and apply them to all compilers, which was exactly what we wanted to avoid...

This PR fixes that by using a variable name different for each test (adding the
language as part of the language name), and uses the code of PR #336 to sanitize
the names of the variables/tests.

It also cleans up the messages, becoming slightly less verbose than before:
it lists the languages that support the flag (or says none) instead of showing
two lines per test.

This PR should supersede PR #336 if it is approved.