Profiling fix: parsec_init(NULL, NULL)
When calling `parsec_init(NULL, NULL)` with profiling enabled and activated,
we would use `app_name_<PID>` as info to create the profile file. However,
when reading a distributed profile, it is assumed that the same info is
passed to all ranks. The manual hinted this through "uniquely" identify
the experiment, but that was not clear enough.

As a result, parallel profile files generated when initializing parsec
with NULL would fail to load.

This patch proposes to pass the same default string (by removing the PID)
to solve this issue.

Alternative approaches could be
  - to add a collective operation here to decide on a unique common
    name
  - to remove the check in dbp_reader.c that this info needs to match

Make the documentation more clear about parsec_profiling_dbp_start

The documentation of this function was stall: it was not updated when
we changed the file naming scheme (removed the mkstemp calll and the
XXXXXX in the filename), and we changed how the rank of the process
is passed to the profiling system. Make it more clear that the string
passed is an identifier, and must match between calling processes.

Implement suggested approach to support the '<app>' profiling filename in single-rank at least