Fallback using a compatible HWLOC.
If the MPI provides HWLOC and it conflicts with the HWLOC version installed by
default on the system, we were not correctly detecting the right HWLOC version
to use. With this patch, if HWLOC_DIR is not specified we will try to use the
MPI provided HWLOC first, if any.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>