profiling tools: more efficient cross-stream event matching
Performance when matching events that start in one profiling stream and
end in a different one is abysmal: O(m * n) in the worst case for m such
events and n total events. This has largely not been a problem, since
the only time this occurs is when using GPUs, and there are relatively
few such events.

In trying to profile my ongoing work with LCI, this is no longer the
case. Communication events almost always start in the progress thread
and end in the communication thread, or vice-versa. This results in many
such events, so the current behavior is infeasible.

This code relies on an observation by George Bosilca that most start
events are immediately followed by their end events. This means it is
possible to construct a list of end events that are possible matches for
cross-stream events. If the matching event for a start event is not
found in its own stream, this list can be searched to determine where to
start the search in other streams. Since the list is naturally sorted by
timestamp, a binary search can be used. I rely on bsearch and a slight
hack that stores the last array element in the search key to implement a
"fuzzy" timestamp search on a per-event-key basis. This list is
constructed lazily when it is needed and only once per stream.

Constructing the cache is O(n).
Search performance should be O(m * log m) for most cases.

There could be some edge cases where the search finds a potential match
in a stream, but the true match is on a different stream; in this case,
the current code still traverses all possibly-matching events in the
mistaken stream. There isn't much to be done about this, as we cannot
know that the matching event is not in the stream until it has been
traversed in its entirety.

Signed-off-by: Omri Mor <omrimor2@illinois.edu>