profiling: disambiguation between certain MPI events
MPI_DATA_PLD_SND and MPI_DATA_PLD_RCV use the dep index as profiling
event ID. Depending on communication ordering, this can result in
mismatches with the end event. Since the combination of sender and tag
are unique, these can be used to form event IDs; the tag is determined
by the sender when it sends the activation message.

Ensure that events use unique IDs: (sender << 32) | (tag + dep_index)

Signed-off-by: Omri Mor <omrimor2@illinois.edu>