Fix current CUDA bug...
Introduction of the NEW optimization in CUDA introduced this bug that
makes the CUDA driver not copy data from RAM to GPU, if the data copy
comes from a direct memory access. The test to detect this is a NEW
tile is wrong in the code, and abusively confuses copies coming
directly from memory with copies coming from a NEW operation.

The proper way to detect this is a NEW data is to check if the
data collection is NULL (meaning it's not a direct access from the data
collection) AND if the repo_entry is NULL (meaning it doesn't have a
predecessor).

Tested with DPLASMA dpotrf on leconte.