Execution stream keeps the highest priority task for local execution.
Instead of pushing the task into the scheduler and then asking for
something to execute, an execution stream will keep the highest priority
tasks as the next ready task. In general this will improve the data
reuse, but it is possible that it will prevent the scheduler to follow
the critical path (because the highest priority task might be less
important that tasks currently available in the scheduler).

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>