Fix statistics management with multiple GPU.
The data_in_array_size was initialized using parsec_nb_devices, but that
variable is not constant until all the devices have been added to a
context. Thus, each module was having statistics arrays with different
sizes. This patch delays the statistics arrays allocation until the end
of the device registration, a moment where we know how many devices are
overall.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>