Fix hash function generation for PTG
To auto-generate the hash function of task classes in PTG, we need
to compute the range of values for a parameter, but this is not
sufficient: we also need the minimum value of that parameter, so
the hash function remains reversible.

In a previous commit, we used the assumption that the leftmost
expression of the range is always the minimum value, but with
local defitions, this is not always the case: in
  m = [im = 0 .. M] %{ return f(im); %}
if f is not monotonic, f(0) is not necessarily the min value for m.

This commit reverts the change that was 'saving' the storage of the
min value. Note that the min value was computed correctly anyway
during the internal init phase, so this was only saving memory.