Variable size data and short protocol
1. Allow sender to send data of any size.
The sender can send less data than expected by the receiver. If we
select the communication protocol based on how much data the receiver
expects, the sender and receiver could diverge on the protocol to
be used (mainly visible when the sent message is below eager, while
the receiver expects more data. So, force the sender to pack the
amount of data per dep, and force the receiver to abide by this amount.

2. bring back short message protocol (aka data embedded directly into the
  activation message
- enable propagation of the data sizes. This is now part of the
  activation message, but unfortunately they will dissapear after the
  first call to get_datatype, and there is no way to retrieve them. The
  side effect of this is that the get_datatype function should only be
  called once, and this is not the case today (the reshape code makes
  heavy usage).
- clean up the reshape code. I don't understand it, it adds a lot of
  overhead on the critical path (a lot of initializations of useless
  structures, and few function calls). I barely tried to minimize the
  overheads, but at some point we need to understand this code and make
  it more user friendly.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>