remote_dep: rotate bcast toplogy around root
The communication topology always prefers to send to lower-ranked nodes
over higher-ranked nodes. This results in a significant communication
imbalance regardless of how a computation is distributed, unless all
communication is point-to-point.

This introduces a pair of functions, remote_dep_rank_to_bit and
remote_dep_bit_to_rank, to translate between rank id 0 .. np-1 and the
bank offset and bit index for the rank_bits array. This translation
"rotates" the ranks around the root of a communication i.e. the source
is always bank 0/bit 0 and there are different topologies depending on
the root of a bcast. If an application's communications are otherwise
balanced then this results in a better communication balance.

Signed-off-by: Omri Mor <omrimor2@illinois.edu>