TLDR; this should address Issue 464
Rollback the terminating step in the local termination detector;
call the termination callback after terminating the taskpool, once the termination detection module does not access it.

During the development of parsec_taskpool_wait (PR 411), we found a race condition that seemed to require this step, but I am unable to remember / re-discover the scenario
where calling the callback before terminating the taskpool matters, and after testing in release and debug many times, I was unable to reproduce. I assume the actual bug
was solved during another commit. As there are use cases where it is useful to call the callback as the last step, I propose to roll this back.