We still need to create the default arena datatype.
In DPLASMA, JDFs do not use the default type, and they use
an external arena_datatype storage to support LAPACK+Tile format
and datatype caching... However that doesn't mean that arena
datatypes can be simply inherited from the data!

Also fix the exit condition to work in distributed for
branching.