Add device async/again support.
In both cases an event is added into the stream and once this event
completes the task is processes according to it's last return: async
tasks being drtopped from the runtime and again tasks being rescheduled
right away on the same stream.

One of the issues with the fact that AGAIN tasks remain on the same
stream is that a certain number of AGAIN tasks (equal to the number
of pending tasks in the stream) on the same stream will block the stream
for new tasks. Thus, in order to provide progress the device tasks
returning AGAIN must be careful with the time spent in the body.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>