co-manager implemented.
The second thread that submits the task to the GPU device is transitioned to a co-manager.
The task is completed by the manager if the co manager has not yet been set.
The task is freed by the manager if it completes the tasks or the task is freed by the
co-manager.