Add a singlification option.
This option allow threads to run on a single physical resource. The
singlification can be done early (for negative values of the MCA parameter) and
will pack the threads on the resources, or late (for positive values of the MCA
parameter) in which case the threads will be spread across the resources.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>