One MPI communicator per AM tag.
Registering all AM on the same communicator drastically increases the
time necessary to MPI for matching, because the order or posted requests
will be variable during the execution, which means that in average the
matching cost will be linear in the number of posted receives. If we
dedicate a communicaotr per tag, then each matching will happen in one
step, as it will match the first requests pending on that communicator
(because a single tag will now go into each communicator).

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>