New communication engine initialization.
Long story short, we are not doing the right thing during the CE
initialization, plus the initialization is split in multiple places
making it extremely hard to follow. This is a first step toward putting
some consistency into that process.
Simplify the code.

Fix the PUT and GET protocols.

When too many PUT and GET are active in same time the communications end
up delayed into the mpi_funnelled_dynamic_req_fifo queue. Unfortunately,
not all information necessary to send the message was saved. This patch
addresses this issue by saving all necessary onesided information.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>