Use pip install --target to test in the compile directory, but pip install --prefix to install final files
pip install --prefix uses different directories depending on the version of Python3 on Linux.
python3.8 was installing things in <prefix>/lib/python-3.8/sites-package (or something close to that)
python3.10 is installing things in <prefix>/local/lib/python-3.10/dist-packages (or something close to that)

pip install --target installs everything in <target>, but does not follow the appropriate hierarchy for an actual install.

With this patch, we use pip install --target python.test to install everything flat *in a known directory* in order to test locally without install.
And we use pip install --prefix to install files with the right hierarchy in an actual install.