Limit the number of recv requests to ensure there is space for sends
Large amounts of GET operations lead to enough recv requests in the
request and the request fifo to prevent any send operation to happen,
causing a deadlock.

Reserve half of the dynamic request array for send requests by limiting
the number of recv requests that can be put into the request array.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>