Add a simple ping-pong test for GPU accelerators
make a token pass from CPU to each GPU, and back, a few times, to check a possible bug found by @devreal.

Part of the DTD interface was not fully ported to HIP

Enable (cuda|hip)_pingpong test in CI

Add a PTG GPU pingpong test to compare with the behavior in DTD -- Work in progress

Tests need to import the appropriate GPU-specific header file, as insert_function_internal.h doesn't do it for them anymore

Enable PTG test over CUDA

Fix errors in data distribution initialization and some DAG errors in the PTG of the GPU pingpong test

Rename files and directories to match the new status of tests (tests/runtime/cuda is renamed tests/runtime/gpu and the pingpong tests are named to specify the API and not a particular device name, since they should work on both GPU types)

Only define the pingpong tests if a suitable compiler is found for the kernels

Do a ping-pong-pong test instead of ping-pong, to see how dependencies are tracked on GPU-to-GPU task dependency

Fix the checks of the pingpong test, and add it in the Testings.cmake

PTG ping-pong test: in order to guide the selection of the best device, the advised data needs to flow from a CPU task, not directly from memory.

Trying to introduce the gpu_nvidia runner in the CI matrix

Add ROCm, create one github_runner-[device].yaml file per device; remove debugging info from CMakeLists.txt

Add some infrastructure to make sure CI does the device tests where it should, and issue an error if things cannot be tested (e.g. because the GPUs are down or the compiler/spack is broken)

Trying to work around the xml2 issue with mesa.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>

Integrate the gpu_amd/release in the test suite

Add support to rocm-smi in check_nb_devices.sh

Conditional CMake command that depends upon the github runner loaded to prepare for testing