dbpreader: missing corner case in cache building
When building the unmatched events cache for a thread, the
code was looking up pairs of event (e1, e2), and checking
if e2 was matching e1, and would add e2 to the cache if e2
is an end not matching e1.

This was overlooking the possibility that the very first
event in the stream is an end event (which wouldn't match the
preceding event since there is no preceding event).

This has a high chance of happening with the async test,
if we run with more threads than the number of asynchronous
tasks, and would trigger test failures on high-core machines
like guyot.