Relative path, symbolic links, and python examples
file(GLOB ...) executes in the source directory, and
PYTHON_TOOLS_BIN_DIR is in the build directory.

As a consequence, when doing
  file(GLOB pyfiles RELATIVE ${PYTHON_TOOLS_BIN_DIR} examples/*)
pyfiles hold values like
  ../../../../../../parsec-src/tools/profiling/python/examples/demo_ptt.py

This works, but if the VPATH is pointed by a symbolic link, at
make install time, ../../../../.. doesn't point to the directory
hosting the parsec source.

I believe the intent was to have
  examples/demo_ptt.py
in pyfiles anyway. To do so, we need to glob it relative to
the CMAKE_CURRENT_SOURCE_DIR (and do the install from the
CMAKE_CURRENT_SOURCE_DIR).