Hotfix: corner case in parsec_dtd_task_snprintf
It's possible that parsec_dtd_task_snprintf is called
when the `->next` field of the parameters copied at
the end of the task are invalid. One should not access
them using `->next` as we do when parsing the user
input, but knowing they are stored as in an array
at the end of the task, starting at byte
`GET_HEAD_OF_PARAM_LIST(dtd_task)`

This bug started to manifest after using the DSL
specific task_snprintf for all task_snprintfs in
PR #603.

This code also correctly manages the case when count_of_params
is 0, but there are no tests like that today.