Fix the futures.
Our initial impementation used an old trick (functions prototypes
without parameters), with the parameters correctly instantiated
by the caller. Unfortunately, this is not supported by newer versions of
the Cstandard, and compilers are becoming more and more picky.
We have to move away from it.

Thus, we now use variadic functions for parsec futures. A quite invasive
PR, but without any logical impact.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>