Minor cleanup of the DTD parameters manipulation. And a proper fix for the dtd_task_snprintf.
The DTA task's parameters were created as an array but also maintained
as a linked list. There was no need for both access modes, so array
stayed while linked list dissapeared.

Fix all the the uses of the DTD task params to be consistent.

This is a more comprehensive replacement for #607, that addresses the
real cause of the problem with DTD task_snprintf reported by
@abouteiller.

Fixes #603.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>