Hotfix: profiling initialized from the wrong thread in comm engine
After we introduced lazy initialization of the comm engine, we left
the profiling initialization for the remote_dep comm engine be
called from the main parsec thread. That initialization must be called
from the communication thread, this patch moves it to the actual
comm thread. (bug found and fixed by @devreal).