Fix the lack of direct GPU to GPU communications in multi-device runs.
In https://github.com/ICLDisco/parsec/pull/570, we moved from using
cuda_index to using device_index in the 'nvlink' mask that decides
if we can directly communicate to another GPU. However, this bitmask
was initialized at query time, before devices get assigned a device_index.
As a consequence, the bitmask was wrong and no direct device to device
communication was happening.

In this PR, we add a step, after all devices have been registered,
to complete this initialization.