Add a function to pre-set infos for all streams/devices/...
Lazy/dynamic allocation of infos can be convenient for the users,
but also detrimental to predictive behaviors. Examples of drawbacks
are when testing memory constrained setups (e.g., when the test uses
zone_malloc to dynamically allocate objects in the info object
constructor, that may fail undeterministically because it depends
how many streams call the constructor), or small scale performance
runs where the number of cublas objects to initialize vary depending
on the run. Calling the new function incurs determinisitic overheads.