Reenable the memory eviction code
While adding support for coroutines and their use of _AGAIN to resubmit
the same task onto the same stream, we have broken the support for using
_AGAIN to trigger the memory evictions when the GPU memory was completely
full.

This PR adds the use of _NEXT as a legitimate return code from the
stream's progress_fct to indicate that the task should be kept into the
stream queues but the error should be escalted upstream. Upon such an
error the higher-level scheduler should trigger the memory eviction
procedure asap

Signed-off-by: George Bosilca <gbosilca@nvidia.com>