List ordered push: search from back if lower than pivot
Determine a pivot point (mid point of priority range) and search from
front or back depending on where the new item lies. Inserting elements
with highest or lowest priority becomes a constant operation and
elements within the current range will on average be O(N/2).

This assumes a linear distribution of priorities. For non-linear
distributions the insertion will be skewed to either front or back but
and the worst case will be O(N) but average complexity would still be
closer to O(N/2) and we win in cases of highest and lowest priorities.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>