Cleanup the __parsec_schedule_vp to make it do what it was supposed to.
Basically, the next_task can only be set if the submission stream is not
NULL and the distance is 0. In all other cases, the tasks will be pushed
on the first execution stream of each VP, with the exception of the
local VP (the VP of the submission execution stream) where the tasks are
pushed locally on the execution stream.

Signed-off-by: George Bosilca <gbosilca@nvidia.com>